add toolbar component to keycube
